requirements.txt
asgiref==3.8.1
bcrypt==4.1.3
Brotli==1.1.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.7
dj-database-url==2.1.0
Django==5.0.6
django-phonenumber-field==7.3.0
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
gunicorn==22.0.0
h11==0.14.0
inflection==0.5.1
packaging==24.0
paramiko==3.4.0
phonenumbers==8.13.36
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyNaCl==1.5.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
uvicorn==0.29.0
whitenoise==6.6.0
